requirements.txt
appnope==0.1.0
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.4.1
certifi==2015.9.6.2
configobj==5.0.6
configparser==3.3.0.post2
dateutils==0.6.6
decorator==4.0.4
facebook-sdk==0.4.0
functools32==3.2.3.post2
hipchat==0.0.1
ipykernel==4.0.3
ipython==4.0.0
ipython-genutils==0.1.0
Jinja2==2.8
jsonschema==2.5.1
jupyter-client==4.0.0
jupyter-core==4.0.6
MarkupSafe==0.23
mistune==0.7.1
nbconvert==4.0.0
nbformat==4.0.0
notebook==4.0.5
path.py==8.1.1
pickleshare==0.5
ptyprocess==0.5
Pygments==2.0.2
python-dateutil==2.4.2
pytz==2015.6
pyzmq==14.7.0
requests==2.7.0
simplegeneric==0.8.1
six==1.9.0
terminado==0.5
tornado==4.2.1
traitlets==4.0.0
wheel==0.24.0
will
